+++ workers/ohsome_quality_analyst/base/indicator.py
@@ -2,7 +2,7 @@
 TODO:
     Describe this module and how to implement child classes
 """
-
+import json
 from abc import ABCMeta, abstractmethod
 from dataclasses import dataclass
 from datetime import datetime, timezone
@@ -14,7 +14,7 @@ from dacite import from_dict
 from geojson import Feature
 
 from ohsome_quality_analyst.utils.definitions import get_layer_definition, get_metadata
-from ohsome_quality_analyst.utils.helper import flatten_dict
+from ohsome_quality_analyst.utils.helper import flatten_dict, json_serialize
 
 
 @dataclass
@@ -43,7 +43,17 @@ class LayerDefinition:
 
 @dataclass
 class Result:
-    """The result of the Indicator."""
+    """The result of the Indicator.
+
+    Attributes:
+        timestamp_oqt (datetime): Timestamp of the creation of the indicator
+        timestamp_osm (datetime): Timestamp of the used OSM data
+            (e.g. Latest timestamp of the ohsome API results)
+        label (str): Traffic lights like quality label
+        value (float): The result value as float ([0, 1])
+        description (str): Description of the result
+        svg (str): Figure of the result as SVG
+    """
 
     timestamp_oqt: datetime
     timestamp_osm: Optional[datetime]
@@ -118,13 +128,21 @@ class BaseIndicator(metaclass=ABCMeta):
 
     @property
     def data(self) -> dict:
-        """All indicator attributes except feature, result, metadata and layer."""
+        """All Indicator object attributes except feature, result, metadata and layer.
+
+        Attributes will be dumped and immediately loaded again by the `json` library.
+        In this process a custom function for serializing data types which are not
+        supported by the `json` library (E.g. numpy datatypes or objects of the
+        `BaseModelStats` class used by the Mapping Saturation Indicator). This will make
+        sure that objects which can be represented as dictionary will be converted to
+        a dictionary.
+        """
         data = vars(self).copy()
         data.pop("result")
         data.pop("metadata")
         data.pop("layer")
         data.pop("feature")
-        return data
+        return json.loads(json.dumps(data, default=json_serialize).encode())
 
     @abstractmethod
     async def preprocess(self) -> None: